feat(game): Added Who's Who game

+++ src/service/game_service.py
@@ -1,4 +1,5 @@
 import asyncio
+from typing import Tuple
 
 from telegram import Update
 from telegram.ext import ContextTypes
@@ -14,8 +15,10 @@ from src.model.error.GroupChatError import GroupChatError, GroupChatException
 from src.model.game.GameOutcome import GameOutcome
 from src.model.game.GameTurn import GameTurn
 from src.model.game.GameType import GameType
+from src.model.game.whoswho.Character import Character
 from src.model.pojo.Keyboard import Keyboard
 from src.service.bounty_service import get_belly_formatted, add_bounty
+from src.service.cron_service import convert_seconds_to_time
 from src.service.message_service import full_message_send, mention_markdown_user, delete_message
 from src.service.notification_service import send_notification
 
@@ -84,8 +87,9 @@ def get_game_authorized_tg_user_ids(game: Game) -> list[int]:
     return [challenger.tg_user_id, opponent.tg_user_id]
 
 
-def get_text(game: Game, game_name: str, is_finished: bool, game_outcome: GameOutcome = None, user_turn: User = None
-             ) -> str:
+def get_text(game: Game, game_name: str, is_finished: bool, game_outcome: GameOutcome = None, user_turn: User = None,
+             is_turn_based: bool = True, character: Character = None, remaining_seconds_to_start: int = None,
+             is_played_in_private_chat: bool = False) -> str:
     """
     Get the text
     :param game: The game object
@@ -93,9 +97,14 @@ def get_text(game: Game, game_name: str, is_finished: bool, game_outcome: GameOu
     :param is_finished: Is the game finished
     :param game_outcome: The game outcome
     :param user_turn: The user turn
+    :param is_turn_based: Is the game turn based
+    :param character: The character, in case of Who's Who
+    :param remaining_seconds_to_start: The remaining seconds to start
+    :param is_played_in_private_chat: Is the game played in private chat
     :return: The text
     """
 
+    added_ot_text = ""
     if is_finished:
         if game_outcome is GameOutcome.CHALLENGER_WON:
             added_ot_text = phrases.GAME_RESULT_WIN.format(mention_markdown_user(game.challenger))
@@ -103,8 +112,17 @@ def get_text(game: Game, game_name: str, is_finished: bool, game_outcome: GameOu
             added_ot_text = phrases.GAME_RESULT_WIN.format(mention_markdown_user(game.opponent))
         else:
             added_ot_text = phrases.GAME_RESULT_DRAW
+
+        if not is_turn_based and character is not None:
+            added_ot_text += phrases.GAME_RESULT_CHARACTER.format(character.get_markdown_mention())
     else:
-        added_ot_text = phrases.GAME_TURN.format(mention_markdown_user(user_turn))
+        if is_turn_based:
+            added_ot_text = phrases.GAME_TURN.format(mention_markdown_user(user_turn))
+        else:
+            if remaining_seconds_to_start is not None:
+                added_ot_text = phrases.GAME_COUNTDOWN.format(convert_seconds_to_time(remaining_seconds_to_start))
+            elif is_played_in_private_chat:
+                added_ot_text = phrases.GAME_STARTED
 
     return phrases.GAME_TEXT.format(game_name,
                                     mention_markdown_user(game.challenger),
@@ -113,17 +131,27 @@ def get_text(game: Game, game_name: str, is_finished: bool, game_outcome: GameOu
                                     added_ot_text)
 
 
-async def delete_game(context: ContextTypes.DEFAULT_TYPE, game: Game, should_delete_message: bool = True) -> None:
+async def delete_game(context: ContextTypes.DEFAULT_TYPE, game: Game, should_delete_message: bool = True,
+                      show_timeout_message: bool = False) -> None:
     """
     Delete game
     :param context: The context
     :param game: The game
     :param should_delete_message: If the message should be deleted
+    :param show_timeout_message: If the message should be edited showing timeout
     :return: None
     """
-    # Try to delete message
-    if should_delete_message:
-        await delete_message(context=context, group_chat=game.group_chat, message_id=game.message_id)
+
+    if should_delete_message and show_timeout_message:
+        raise ValueError("should_delete_message and show_timeout_message cannot be both True")
+
+    if show_timeout_message:
+        await full_message_send(context=context, group_chat=game.group_chat, text=phrases.GAME_TIMEOUT,
+                                edit_message_id=game.message_id)
+    elif should_delete_message:
+        # Try to delete message
+        if should_delete_message:
+            await delete_message(context=context, group_chat=game.group_chat, message_id=game.message_id)
 
     # Return wager to challenger
     challenger: User = game.challenger
@@ -226,5 +254,37 @@ def get_game_name(game_type: GameType) -> str:
         case GameType.RUSSIAN_ROULETTE:
             return phrases.RUSSIAN_ROULETTE_GAME_NAME
 
+        case GameType.WHOS_WHO:
+            return phrases.WHOS_WHO_GAME_NAME
+
         case _:
             return phrases.GAME_UNKNOWN_NAME
+
+
+async def enqueue_game_timeout(context: ContextTypes.DEFAULT_TYPE, game: Game):
+    """
+    Enqueue a game timeout. Waits for N time and if the opponent doesn't accept, the game is deleted
+    :param context: The context
+    :param game: The game
+    :return: None
+    """
+
+    # Wait for N time
+    await asyncio.sleep(Env.GAME_CONFIRMATION_TIMEOUT.get_int())
+
+    updated_game = Game.get_by_id(game.id)
+
+    # Check if the game is still in the same state
+    if GameStatus(updated_game.status) == GameStatus.AWAITING_OPPONENT_CONFIRMATION:
+        await delete_game(context, updated_game, should_delete_message=False, show_timeout_message=True)
+
+
+def get_participants(game: Game) -> Tuple[User, User]:
+    """
+    Get the participants of a game
+
+    :param game: The game
+    :return: The participants
+    """
+
+    return game.challenger, game.opponent
